Handle main entry point by SDL.
git-svn-id: https://cvet.by:8443/svn/FOnlineSource@534 c2c6a35e-ad2b-4e1f-bb64-97b989be360b

Former-commit-id: 8984c02c472091831348757befd0ee8a7ab924f6

+++ Source/ClientInterface.cpp
@@ -6367,7 +6367,7 @@ void FOClient::GmapLMouseDown()
     }
     else
     {
-        for( int i = 0; i < HexMngr.GetCritters().size(); i++ )
+        for( int i = 0; i < (int) HexMngr.GetCritters().size(); i++ )
         {
             if( !IsCurInRect( Rect( GmapWName, GmapWNameStepX * i, GmapWNameStepY * i ) ) )
                 continue;
